Moved Strategy related code to own file
